worker/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
openai==1.12.0
qdrant-client==1.7.0
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-dotenv==1.0.0
tiktoken==0.5.2
numpy==1.24.3
pytest==7.4.3
pytest-asyncio==0.21.1
beautifulsoup4==4.12.2
requests==2.31.0
httpx==0.25.2
firecrawl-py==0.0.20
